Move: move model contracts into contracts/user.
