+++ packages/admin/src/Contracts/User/HasName.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Filament\Contracts\User;
+
+interface HasName
+{
+    /**
+     * Returns the name that is displayed in filament.
+     *
+     * @return string
+     */
+    public function getFilamentName(): string;
+}
