--- packages/admin/src/Models/Contracts/FilamentUser.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Filament\Models\Contracts;
-
-interface FilamentUser
-{
-    public function canAccessFilament(): bool;
-}
