--- packages/admin/src/Models/Contracts/HasAvatar.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Filament\Models\Contracts;
-
-interface HasAvatar
-{
-    public function getFilamentAvatarUrl(): ?string;
-}
