--- packages/admin/src/Models/Contracts/HasName.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Filament\Models\Contracts;
-
-interface HasName
-{
-    public function getFilamentName(): string;
-}
